requirements.txt
# BLOOMCRAWLER RIIS - Python Dependencies
# Install with: pip install -r requirements.txt

# Core Web Framework
Flask==3.0.0
flask-socketio==5.3.6
python-socketio==5.8.0

# GUI Framework
tkinter  # Usually comes with Python

# Data Science & ML
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Visualization
matplotlib==3.7.2
plotly==5.15.0
networkx==3.1

# System Monitoring
psutil==5.9.5
GPUtil==1.4.0

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2

# Additional Utilities
asyncio-mqtt==0.13.1
eventlet==0.33.3
